Apportate modifiche a Note.java, NoteEmotion.java, Emotion.java e sistemato Flyway. Aggiunto plugin di Flyway per poter usare mvn flyway:repair

+++ src/main/resources/db/migration/V7__create_ai_analysis.sql
@@ -4,7 +4,7 @@ CREATE TABLE ai_analysis (
 
     note_id BIGINT NOT NULL UNIQUE,
 
-    mood_score INTEGER,
+    mood_score DOUBLE PRECISION,
 
     summary TEXT,
